# EcommerceOrderMS/Ecommerce.OrderMS.Domain/Model/OrderItemUpdateRequest.cs
namespace Ecommerce.OrderMS.Domain.Model;

public record OrderItemUpdateRequest(Guid ProductID, decimal UnitPrice, int Quantity)
{
  public OrderItemUpdateRequest() : this(default, default, default)
  {
  }
}